+++ spring-boot-project/spring-boot/src/main/java/org/springframework/boot/web/embedded/jetty/JettyReactiveWebServerFactory.java
@@ -97,11 +97,13 @@ public class JettyReactiveWebServerFactory extends AbstractReactiveWebServerFact
 		ServletContextHandler contextHandler = new ServletContextHandler(server, "",
 				false, false);
 		contextHandler.addServlet(servletHolder, "/");
+		this.logger.info("Server initialized with port: " + port);
+		SslServerCustomizer sslServerCustomizer = new SslServerCustomizer(port,
+				getSsl(), getSslStoreProvider());
+		sslServerCustomizer.customize(server);
 		for (JettyServerCustomizer customizer : getServerCustomizers()) {
 			customizer.customize(server);
 		}
-		JettyReactiveWebServerFactory.logger
-				.info("Server initialized with port: " + port);
 		return server;
 	}
