+++ spring-boot-project/spring-boot/src/main/java/org/springframework/boot/web/embedded/netty/NettyReactiveWebServerFactory.java
@@ -90,6 +90,10 @@ public class NettyReactiveWebServerFactory extends AbstractReactiveWebServerFact
 	private HttpServer createHttpServer() {
 		return HttpServer.builder().options((options) -> {
 			options.listenAddress(getListenAddress());
+			if (getSsl() != null && getSsl().isEnabled()) {
+				SslServerCustomizer sslServerCustomizer = new SslServerCustomizer(getSsl(), getSslStoreProvider());
+				sslServerCustomizer.customize(options);
+			}
 			applyCustomizers(options);
 		}).build();
 	}
